split out CSTextStyle
